Refactor app into modular structure
