+++ composer.json
@@ -1,0 +1,13 @@
+{
+    "name": "wish/wall",
+    "description": "Wish Wall app",
+    "type": "project",
+    "require": {
+        "php": "^8.3"
+    },
+    "autoload": {
+        "psr-4": {
+            "WishWall\\": "src/"
+        }
+    }
+}
