+++ src/Router.php
@@ -1,0 +1,30 @@
+<?php
+
+declare(strict_types=1);
+
+namespace WishWall;
+
+final class Router
+{
+    /** @var array<string, array<string, callable>> */
+    private array $routes = [];
+
+    public function add(string $method, string $path, callable $handler): void
+    {
+        $method = strtoupper($method);
+        $this->routes[$method][$path] = $handler;
+    }
+
+    public function dispatch(string $method, string $path): void
+    {
+        $method = strtoupper($method);
+        $handler = $this->routes[$method][$path] ?? null;
+        if ($handler === null) {
+            http_response_code(404);
+            echo 'Not Found';
+            return;
+        }
+        $handler();
+    }
+}
+
